Restructure to split engine, players, and app.
