+++ id_expl0rer/engine/CMakeLists.txt
@@ -9,8 +9,9 @@ add_library(${PROJECT_NAME} src/engine.cpp)
 target_include_directories(${PROJECT_NAME}
         PUBLIC
             $<INSTALL_INTERFACE:include>
-            $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include
+            $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
         PRIVATE
+            ${CMAKE_CURRENT_SOURCE_DIR}/src
         )
 
 target_link_libraries(${PROJECT_NAME}
@@ -19,3 +20,5 @@ target_link_libraries(${PROJECT_NAME}
         PRIVATE
 
         )
+
+add_subdirectory(players/desktop players/wristterm)
