--- id_expl0rer/engine/include/engine.h
@@ -1,5 +1,0 @@
-#pragma once
-
-class id_expl0rerEngine {
-    void PrintHelloWorld();
-};
