+++ id_expl0rer/engine/include/id_expl0rer_engine/engine.h
@@ -1,0 +1,6 @@
+#pragma once
+
+class id_expl0rerEngine {
+    public:
+        static void PrintHelloWorld();
+};
