+++ id_expl0rer/engine/players/wristterm/gwes-sample2.vcproj
