+++ id_expl0rer/engine/players/wristterm/src/gwessample2ppc.rc2
