+++ id_expl0rer/engine/players/wristterm/src/resourceppc.h
