+++ id_expl0rer/engine/tests/CMakeLists.txt
@@ -1,0 +1,7 @@
+cmake_minimum_required(VERSION 3.5)
+
+project(id_expl0rer_tests)
+
+set(CMAKE_CXX_STANDARD 98)
+
+find_package(Boost COMPONENTS unit_test_framework REQUIRED)
